Interrupt worlking
using schmit trigger

+++ OMO_ADUINO_IDE/types.h
@@ -50,8 +50,8 @@ typedef unsigned short uint16;
 typedef signed short int16;
 typedef unsigned int uint32;
 typedef signed int int32;
-typedef unsigned long long uint64;
-typedef signed long long int64;
+typedef unsigned long uint64;
+typedef signed long int64;
 typedef float fint32;
 
 typedef struct
